refactor dubbo-samples-version
