+++ 2-advanced/dubbo-samples-version/dubbo-samples-version-consumer/src/main/resources/log4j.properties
