--- 2-advanced/dubbo-samples-version/src/main/java/org/apache/dubbo/samples/version/VersionConsumer.java
@@ -1,39 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-
-package org.apache.dubbo.samples.version;
-
-import org.apache.dubbo.samples.version.api.VersionService;
-
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-public class VersionConsumer {
-
-    public static void main(String[] args) throws Exception {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("spring/version-consumer.xml");
-        context.start();
-
-        VersionService versionService = (VersionService) context.getBean("versionService");
-        for (int i = 0; i < 10000; i++) {
-            String hello = versionService.sayHello("world");
-            System.out.println(hello);
-            Thread.sleep(2000);
-        }
-    }
-}
