--- 2-advanced/dubbo-samples-version/src/test/java/org/apache/dubbo/samples/version/MyNettyTransporter.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.samples.version;
-
-import org.apache.dubbo.common.URL;
-import org.apache.dubbo.remoting.ChannelHandler;
-import org.apache.dubbo.remoting.Client;
-import org.apache.dubbo.remoting.RemotingException;
-import org.apache.dubbo.remoting.transport.netty4.NettyTransporter;
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-public class MyNettyTransporter extends NettyTransporter {
-
-    public static final String NAME = "myNetty";
-
-    private static AtomicInteger connectedCount = new AtomicInteger(0);
-
-    @Override
-    public Client connect(URL url, ChannelHandler handler) throws RemotingException {
-        Client client = super.connect(url, handler);
-        if (client.isConnected()) {
-            connectedCount.incrementAndGet();
-        }
-        return client;
-    }
-
-    public static int getConnectedCount() {
-        return connectedCount.get();
-    }
-    
-    public static void reset() {
-        connectedCount.set(0);
-    }
-}
