--- 2-advanced/dubbo-samples-version/src/test/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.AddressListener
@@ -1,1 +1,0 @@
-test=org.apache.dubbo.samples.version.MyAddressListener
