--- 2-advanced/dubbo-samples-version/src/test/resources/META-INF/dubbo/internal/org.apache.dubbo.remoting.Transporter
@@ -1,1 +1,0 @@
-myNetty=org.apache.dubbo.samples.version.MyNettyTransporter
